Refactor server structure and improve path handling
Moved server files to more logical directories, updated path handling for certs and public assets, and consolidated BYOND communication logic. Removed redundant files and improved parent process monitoring. Updated README with clearer build instructions and architecture notes.
