+++ voicechat/node/package.json
@@ -1,7 +1,7 @@
 {
   "name": "signaling-server",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "server/main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
