+++ voicechat/node/server/byond/ByondServer.js
@@ -1,7 +1,20 @@
+const fs = require('fs');
 const net = require('net');
-const { sendJSON } = require('./byondCommunication');
+const path = require('path');
+
+const { sendJSON } = require('./ByondCommunication');
 const { handleRequest } = require('./ByondHandlers');
-const PIPE_NAME = '/tmp/byond_node.sock';
+
+const PIPE_NAME = 'byond_node.sock';
+const PIPE_PATH = path.resolve(process.cwd(), PIPE_NAME)
+
+function cleanUpExistingSocket(PIPE_PATH){
+    if(fs.existsSync(PIPE_PATH)){
+        console.log(`found existing named socket at ${PIPE_NAME} removing...`)
+        fs.unlinkSync(PIPE_PATH);
+    }
+}
+
 function startByondServer(byondPort, io, shutdown_function) {
     const ByondServer = net.createServer((stream) => {
         stream.on('data', (data) => {
@@ -19,8 +32,8 @@ function startByondServer(byondPort, io, shutdown_function) {
         stream.on('end', () => {
         });
     });
-
-    ByondServer.listen(PIPE_NAME, () => {
+    cleanUpExistingSocket(PIPE_PATH)
+    ByondServer.listen(PIPE_PATH, () => {
         console.log(`socket server listening on ${PIPE_NAME}`);
     });
